# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Beauty Clinic Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      parameters:
        - name: date
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: patientId
          in: query
          schema: { type: integer }
        - name: staffId
          in: query
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetailsList"
    post:
      operationId: createAppointment
      tags: [appointments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
...
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
        "404":
          description: Not found
    put:
      operationId: updateAppointment
      tags: [appointments]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDetails"
        "404":
          description: Not found
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Payments ───────────────────────────────────────────────────────────────
...
  /commission-recipients:
    get:
      operationId: listCommissionRecipients
      tags: [commission-recipients]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommissionRecipient"
    post:
      operationId: createCommissionRecipient
      tags: [commission-recipients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommissionRecipientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionRecipient"
  /commission-recipients/{id}/referrals:
    get:
      operationId: getCommissionRecipientReferrals
      tags: [commission-recipients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionRecipientReferrals"
  /commission-recipients/{id}:
    put:
      operationId: updateCommissionRecipient
      tags: [commission-recipients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommissionRecipientUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionRecipient"
        "404":
          description: Not found
    delete:
      operationId: deleteCommissionRecipient
      tags: [commission-recipients]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Reminders ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AppointmentWithDetails:
      type: object
      required: [id, patientId, serviceId, scheduledAt, status, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        serviceId: { type: integer }
        staffId: { type: ["integer", "null"] }
        scheduledAt: { type: integer }
        status: { type: string }
        notes: { type: ["string", "null"] }
        price: { type: ["integer", "null"] }
        discountId: { type: ["integer", "null"] }
        originalPrice: { type: ["integer", "null"] }
        deposit: { type: ["integer", "null"] }
        createdAt: { type: integer }
        patientName: { type: ["string", "null"] }
        patientPhone: { type: ["string", "null"] }
        patientFileNumber: { type: ["string", "null"] }
        patientTier: { type: ["string", "null"] }
        serviceName: { type: ["string", "null"] }
        servicePrice: { type: ["integer", "null"] }
        staffName: { type: ["string", "null"] }
        unitsUsed: { type: ["integer", "null"] }
        priceMode: { type: ["string", "null"] }
        unitPrice: { type: ["integer", "null"] }
        unitLabel: { type: ["string", "null"] }
        serviceUnitCount: { type: ["integer", "null"] }
    AppointmentWithDetailsList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentWithDetails"
        total: { type: integer }
    AppointmentInput:
      type: object
      required: [patientId, serviceId, scheduledAt]
      properties:
        patientId: { type: integer }
        serviceId: { type: integer }
        staffId: { type: integer }
        scheduledAt: { type: integer }
        status: { type: string }
        notes: { type: string }
        price: { type: integer }
        discountId: { type: integer }
        originalPrice: { type: integer }
        deposit: { type: integer }
    AppointmentUpdate:
      type: object
      properties:
        patientId: { type: integer }
        serviceId: { type: integer }
        staffId: { type: ["integer", "null"] }
        scheduledAt: { type: integer }
        status: { type: string }
        notes: { type: ["string", "null"] }
        price: { type: ["integer", "null"] }
        discountId: { type: ["integer", "null"] }
        originalPrice: { type: ["integer", "null"] }
        unitsUsed: { type: ["integer", "null"] }

    # ── Payment ───────────────────────────────────────────────────────────────
...
    CommissionRecipientReferral:
      type: object
      required: [patientId, name, totalSpent, commission]
      properties:
        patientId: { type: integer }
        name: { type: string }
        fileNumber: { type: ["string", "null"] }
        totalSpent: { type: integer }
        referrerRate: { type: ["integer", "null"] }
        commission: { type: integer }
    CommissionRecipientReferrals:
      type: object
      required: [recipient, count, totalSpent, totalCommission, referrals]
      properties:
        recipient:
          type: object
          required: [id, name]
          properties:
            id: { type: integer }
            name: { type: string }
        count: { type: integer }
        totalSpent: { type: integer }
        totalCommission: { type: integer }
        referrals:
          type: array
          items:
            $ref: "#/components/schemas/CommissionRecipientReferral"
    CommissionRecipient:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        createdAt: { type: integer }
    CommissionRecipientInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        phone: { type: string }
        description: { type: string }
    CommissionRecipientUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: ["string", "null"] }
        description: { type: ["string", "null"] }

    # ── Commission ────────────────────────────────────────────────────────────
...